adds instrument panel functionality and refactoring
