+++ Assets/Resources/Materials/Cloak.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4e049bd0ef2ea51428a4d133d430e6aa
+timeCreated: 1465252563
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
