+++ Assets/Scripts/ActivateSongAction.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 7d42fa01f590322419daeaecbab58fbb
+timeCreated: 1465272262
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
