+++ Assets/Scripts/MuteSoloAction.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 095f408c46da8c54ab0df47880d356f7
+timeCreated: 1465269741
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
