+++ Assets/Scripts/MuteSoloController.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 1d13ccf0fff6ea245ba98abdff48a668
+timeCreated: 1465255261
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
